removed logo flag; added -l --luxtorpeda update flag; added -u --unsafe flag to omit sha512sum check

+++ scripts/utils/update-proton-ge.py
@@ -29,11 +29,25 @@ if utils.is_root():
 
 DOWNLOAD_DIR = "/tmp/"
 PROTON_GE_GITHUB_RELEASES_URL = "https://api.github.com/repos/GloriousEggroll/proton-ge-custom/releases"
+LUXTORPEDA_GITHUB_RELEASES_URL = "https://api.github.com/repos/luxtorpeda-dev/luxtorpeda/releases"
+COMPATIBILITY_LAYER_URL = PROTON_GE_GITHUB_RELEASES_URL # default compat layer to install 
 INSTALL_DIR = os.path.expanduser("~/.local/share/Steam/compatibilitytools.d/")
 VERSION = None
 
 parser = ap.ArgumentParser(
     description="Download & extract latest version of GE-Proton\n\thttps://github.com/GloriousEggroll/proton-ge-custom"
+)
+parser.add_argument(
+    "-l", "--luxtorpeda", 
+    help="Download & extract latest version of Luxtorpeda\n\thttps://github.com/luxtorpeda-dev/luxtorpeda", 
+    required=False, 
+    action="store_true"
+)
+parser.add_argument(
+    "-u", "--unsafe",
+    help="do not check (verify) hash download contents with sha512sum",
+    required=False,
+    action="store_true"
 )
 parser.add_argument("-d", "--destination", help="specify installation directory", required=False)
 parser.add_argument("-t", "--temporary", help="specify temporary download directory", required=False)
@@ -43,10 +57,8 @@ parser.add_argument(
     required=False,
     action="store_true"
 )
-parser.add_argument("-l", "--logo", help="do not print the utility logo", required=False, action="store_true")
 parser.add_argument(
-    "-v",
-    "--version",
+    "-v", "--version",
     help="specific version to install, with standard GE-Proton naming format e.g. 7-46",
     required=False,
     default=None
@@ -67,6 +79,10 @@ if args.temporary:
 if args.version:
     VERSION = args.version
 
+if args.luxtorpeda:
+    args.unsafe = True # as of writing, luxtorpeda doesn't have a sha512sum in their assets.
+    COMPATIBILITY_LAYER_URL = LUXTORPEDA_GITHUB_RELEASES_URL
+
 if args.subcommand:
     match args.subcommand:
         case "ls":
@@ -74,68 +90,74 @@ if args.subcommand:
             for d in dirs:
                 print(d)
             exit(0)
-        case "versions":
-            for v in utils.get_github_releases(PROTON_GE_GITHUB_RELEASES_URL, recurse=True):
+        case "versions":            
+            for v in utils.get_github_releases(COMPATIBILITY_LAYER_URL, recurse=True):
                 print(v.tag_name)
             exit(0)
         case _:
             print("Unknown subcommand, exiting...")
             exit(2)
 
-if not args.logo:
-    print("""
-    \033[5m
-     _____  _____                   _              
-    |  __ \|  ___|                 | |             
-    | |  \/| |__    _ __  _ __ ___ | |_ ___  _ __  
-    | | __ |  __|  | '_ \| '__/ _ \| __/ _ \| '_ \ 
-    | |_\ \| |___  | |_) | | | (_) | || (_) | | | |
-     \____/\____/  | .__/|_|  \___/ \__\___/|_| |_|
-             ______| |______ ______ ______ ______  
-        _   |______|_|______|______|______|______| 
-       (_)         | |      | | |                  
-        _ _ __  ___| |_ __ _| | | ___ _ __         
-       | | '_ \/ __| __/ _` | | |/ _ \ '__|        
-       | | | | \__ \ || (_| | | |  __/ |           
-       |_|_| |_|___/\__\__,_|_|_|\___|_|           
-    \033[0m
-    """)
+print("""
+\033[5m
+ _____  _____                   _              
+|  __ \|  ___|                 | |             
+| |  \/| |__    _ __  _ __ ___ | |_ ___  _ __  
+| | __ |  __|  | '_ \| '__/ _ \| __/ _ \| '_ \ 
+| |_\ \| |___  | |_) | | | (_) | || (_) | | | |
+ \____/\____/  | .__/|_|  \___/ \__\___/|_| |_|
+         ______| |______ ______ ______ ______  
+    _   |______|_|______|______|______|______| 
+   (_)         | |      | | |                  
+    _ _ __  ___| |_ __ _| | | ___ _ __         
+   | | '_ \/ __| __/ _` | | |/ _ \ '__|        
+   | | | | \__ \ || (_| | | |  __/ |           
+   |_|_| |_|___/\__\__,_|_|_|\___|_|           
+\033[0m
+""")
 
 if not os.path.exists(INSTALL_DIR):
     os.makedirs(INSTALL_DIR)
 
-release = match_correct_release(PROTON_GE_GITHUB_RELEASES_URL, title=VERSION)
+release = match_correct_release(COMPATIBILITY_LAYER_URL, title=VERSION)
 if not release:
     print(f"Couldn't match any release for version {VERSION}")
     exit(1)
-print(f"Found correct version {release.tag_name}")
+print(f"Found correct version{' Luxtorpeda' if args.luxtorpeda else '' } {release.tag_name}")
+
 ftarballname = None
 ftarballurl = None
 fhashname = None
 fhashurl = None
 for n, l in release.assets.items():
-    if "tar.gz" in n:
+    if "tar" in n:
         ftarballname = DOWNLOAD_DIR + os.sep + n
         ftarballurl = l
     if "sha512sum" in n:
         fhashname = DOWNLOAD_DIR + os.sep + n
         fhashurl = l
 
-if not fhashname or not ftarballname:
-    print(f"Couldn't find a sha512sum or tarball for version {release.tag_name}")
+if not fhashname:
+    print(f"Couldn't find a sha512sum for version {release.tag_name}")
+    if not args.unsafe:
+        exit(1)
+    
+if not ftarballname:
+    print(f"Couldn't find a tarball for version {release.tag_name}")
     exit(1)
 
 # download sha512sum
-try:
-    with open(fhashname, "wb") as out:
-        print(f"Writing {fhashname} from url {fhashurl}")
-        for bread, btotal, data in utils.download(fhashurl):
-            out.write(data)
-            utils.echo_progress_bar_complex(bread, btotal, sys.stdout, os.get_terminal_size().columns)
-    print(f"\nDownloaded {fhashname}")
-except Exception as e:
-    print(e, out=sys.stderr)
-    exit(1)
+if not args.unsafe:
+    try:
+        with open(fhashname, "wb") as out:
+            print(f"Writing {fhashname} from url {fhashurl}")
+            for bread, btotal, data in utils.download(fhashurl):
+                out.write(data)
+                utils.echo_progress_bar_complex(bread, btotal, sys.stdout, os.get_terminal_size().columns)
+        print(f"\nDownloaded {fhashname}")
+    except Exception as e:
+        print(e, out=sys.stderr)
+        exit(1)
 
 # download tarball
 try:
@@ -149,17 +171,20 @@ except Exception as e:
     print(e, out=sys.stderr)
     exit(1)
 
-if not utils.run_subprocess(["sha512sum", "-c", fhashname], DOWNLOAD_DIR):
-    if not args.keep:
-        os.remove(fhashname)
-        os.remove(ftarballname)
-    exit(1)
+# there's no sha512sum to download in luxtorpeda (as of writing)
+if not args.unsafe:
+    if not utils.run_subprocess(["sha512sum", "-c", fhashname], DOWNLOAD_DIR):
+        if not args.keep:
+            os.remove(fhashname)
+            os.remove(ftarballname)
+        exit(1)
 
 time.sleep(1)
 
 if not utils.run_subprocess(["tar", "-xPf", ftarballname, f"--directory={INSTALL_DIR}"], DOWNLOAD_DIR):
     if not args.keep:
-        os.remove(fhashname)
+        if not args.unsafe:
+            os.remove(fhashname)
         os.remove(ftarballname)
     exit(1)
 print(f"Extracted {ftarballname}")
@@ -173,8 +198,9 @@ print(f"Extracted {ftarballname}")
 
 if not args.keep:
     os.remove(ftarballname)
-    os.remove(fhashname)
-    print(f"Removed {DOWNLOAD_DIR} files {ftarballname} & {fhashname}")
+    if not args.unsafe:
+        os.remove(fhashname)
+    print(f"Removed {DOWNLOAD_DIR} files.")
 
 print(f"New contents are at {INSTALL_DIR}")
 print("Done.")
